# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Casenova API Documentation",
        "description": "Casenova API Documentation",
        "contact": {
            "email": "admin@casenova.uz"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://casenova.lc/api/v1",
            "description": "Api Server"
        }
    ],
    "paths": {
...
        "/services": {
            "get": {
                "tags": [
                    "Services"
                ],
                "summary": "Get list of services",
                "description": "Returns list of services",
                "operationId": "getServices",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ServiceCollection"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            },
            "post": {
                "tags": [
                    "Services"
                ],
                "summary": "Create new service",
                "description": "Returns service data",
                "operationId": "storeServices",
                "requestBody": {
                    "description": "Pass user credentials",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/StoreServiceRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Successful operation"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            }
        },
        "/services/{id}": {
            "get": {
                "tags": [
                    "Services"
                ],
                "summary": "Get service information",
                "description": "Returns service data",
                "operationId": "getServiceById",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Service id",
                        "required": true,
                        "schema": {
                            "$ref": "#/components/schemas/BaseModel/properties/property_id"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/ServiceResource"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            },
            "put": {
                "tags": [
                    "Services"
                ],
                "summary": "Update service",
                "description": "Returns service data",
                "operationId": "updateServices",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Service id",
                        "required": true,
                        "schema": {
                            "$ref": "#/components/schemas/BaseModel/properties/property_id"
                        }
                    }
                ],
                "requestBody": {
                    "description": "Pass user credentials",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdateServiceRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Successful operation"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            }
        },
...
        "/subservices": {
            "get": {
                "tags": [
                    "Subservices"
                ],
                "summary": "Get list of subservices",
                "description": "Returns list of subservices",
                "operationId": "getSubservices",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/SubserviceCollection"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            },
            "post": {
                "tags": [
                    "Subservices"
                ],
                "summary": "Create new subservice",
                "description": "Returns service data",
                "operationId": "storeSubservices",
                "requestBody": {
                    "description": "Pass user credentials",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/StoreSubserviceRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Successful operation"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            }
        },
        "/subservices/{id}": {
            "get": {
                "tags": [
                    "Services"
                ],
                "summary": "Get subservice information",
                "description": "Returns subservice data",
                "operationId": "getSubserviceById",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Subservice id",
                        "required": true,
                        "schema": {
                            "$ref": "#/components/schemas/BaseModel/properties/property_id"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/SubserviceResource"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            },
            "put": {
                "tags": [
                    "Subservices"
                ],
                "summary": "Update subservice",
                "description": "Returns subservice data",
                "operationId": "updateSubservice",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Subservice id",
                        "required": true,
                        "schema": {
                            "$ref": "#/components/schemas/BaseModel/properties/property_id"
                        }
                    }
                ],
                "requestBody": {
                    "description": "Pass user credentials",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdateSubserviceRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Successful operation"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            }
        },
        "/subservices/{subservice}/image": {
            "post": {
                "tags": [
                    "Subservices"
                ],
                "summary": "Update image for subservice",
                "description": "Returns subservice status",
                "operationId": "imageSubservice",
                "parameters": [
                    {
                        "name": "subservice",
                        "in": "path",
                        "description": "Subservice id",
                        "required": true,
                        "schema": {
                            "$ref": "#/components/schemas/BaseModel/properties/property_id"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "properties": {
                                    "image": {
                                        "$ref": "#/components/schemas/BaseModel/properties/property_uploads"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/BaseModel/properties/property_boolean_result"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "StoreServiceRequest": {
                "title": "StoreServiceRequest",
                "properties": {
                    "ru": {
                        "$ref": "#/components/schemas/StoreServiceTranslationRequest"
                    },
                    "uz": {
                        "$ref": "#/components/schemas/StoreServiceTranslationRequest"
                    },
                    "en": {
                        "$ref": "#/components/schemas/StoreServiceTranslationRequest"
                    }
                },
                "type": "object",
                "xml": {
                    "name": "StoreServiceRequest"
                }
            },
            "StoreServiceTranslationRequest": {
                "title": "StoreServiceTranslationRequest",
                "properties": {
                    "name": {
                        "$ref": "#/components/schemas/ServiceTranslation/properties/name"
                    },
                    "title": {
                        "$ref": "#/components/schemas/ServiceTranslation/properties/title"
                    }
                },
                "type": "object",
                "xml": {
                    "name": "StoreServiceTranslationRequest"
                }
            },
            "StoreSubserviceRequest": {
                "title": "StoreSubserviceRequest",
                "properties": {
                    "ru": {
                        "$ref": "#/components/schemas/StoreSubserviceTranslationRequest"
                    },
                    "uz": {
                        "$ref": "#/components/schemas/StoreSubserviceTranslationRequest"
                    },
                    "en": {
                        "$ref": "#/components/schemas/StoreSubserviceTranslationRequest"
                    }
                },
                "type": "object",
                "xml": {
                    "name": "StoreSubserviceRequest"
                }
            },
            "StoreSubserviceTranslationRequest": {
                "title": "StoreSubserviceTranslationRequest",
                "properties": {
                    "name": {
                        "$ref": "#/components/schemas/SubserviceTranslation/properties/name"
                    },
                    "content": {
                        "$ref": "#/components/schemas/SubserviceTranslation/properties/content"
                    }
                },
                "type": "object",
                "xml": {
                    "name": "StoreSubserviceTranslationRequest"
                }
            },
...
            "UpdateServiceRequest": {
                "title": "UpdateServiceRequest",
                "properties": {
                    "ru": {
                        "$ref": "#/components/schemas/UpdateServiceTranslationRequest"
                    },
                    "uz": {
                        "$ref": "#/components/schemas/UpdateServiceTranslationRequest"
                    },
                    "en": {
                        "$ref": "#/components/schemas/UpdateServiceTranslationRequest"
                    }
                },
                "type": "object",
                "xml": {
                    "name": "UpdateServiceRequest"
                }
            },
            "UpdateServiceTranslationRequest": {
                "title": "UpdateServiceTranslationRequest",
                "properties": {
                    "name": {
                        "$ref": "#/components/schemas/ServiceTranslation/properties/name"
                    },
                    "title": {
                        "$ref": "#/components/schemas/ServiceTranslation/properties/title"
                    }
                },
                "type": "object",
                "xml": {
                    "name": "UpdateServiceTranslationRequest"
                }
            },
            "UpdateSubserviceRequest": {
                "title": "UpdateSubserviceRequest",
                "properties": {
                    "ru": {
                        "$ref": "#/components/schemas/UpdateSubserviceTranslationRequest"
                    },
                    "uz": {
                        "$ref": "#/components/schemas/UpdateSubserviceTranslationRequest"
                    },
                    "en": {
                        "$ref": "#/components/schemas/UpdateSubserviceTranslationRequest"
                    }
                },
                "type": "object",
                "xml": {
                    "name": "UpdateSubserviceRequest"
                }
            },
            "UpdateSubserviceTranslationRequest": {
                "title": "UpdateSubserviceTranslationRequest",
                "properties": {
                    "name": {
                        "$ref": "#/components/schemas/SubserviceTranslation/properties/name"
                    },
                    "content": {
                        "$ref": "#/components/schemas/SubserviceTranslation/properties/content"
                    }
                },
                "type": "object",
                "xml": {
                    "name": "UpdateSubserviceTranslationRequest"
                }
            },
...
            "ServiceCollection": {
                "title": "ServiceCollection",
                "properties": {
                    "data": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/ServiceResource"
                        }
                    }
                },
                "type": "object"
            },
            "ServiceResource": {
                "title": "ServiceResource",
                "properties": {
                    "id": {
                        "$ref": "#/components/schemas/Service/properties/id"
                    },
                    "image": {
                        "$ref": "#/components/schemas/Service/properties/image"
                    },
                    "createdAt": {
                        "$ref": "#/components/schemas/Service/properties/created_at"
                    },
                    "updatedAt": {
                        "$ref": "#/components/schemas/Service/properties/updated_at"
                    },
                    "translations": {
                        "type": "array",
                        "items": {
                            "anyOf": [
                                {
                                    "$ref": "#/components/schemas/ServiceTranslationResource"
                                }
                            ]
                        }
                    }
                },
                "type": "object",
                "xml": {
                    "name": "ServiceResource"
                }
            },
...
            "ServiceTranslationResource": {
                "title": "ServiceTranslationResource",
                "properties": {
                    "id": {
                        "$ref": "#/components/schemas/ServiceTranslation/properties/id"
                    },
                    "serviceId": {
                        "$ref": "#/components/schemas/ServiceTranslation/properties/service_id"
                    },
                    "locale": {
                        "$ref": "#/components/schemas/ServiceTranslation/properties/locale"
                    },
                    "name": {
                        "$ref": "#/components/schemas/ServiceTranslation/properties/name"
                    },
                    "title": {
                        "$ref": "#/components/schemas/ServiceTranslation/properties/title"
                    }
                },
                "type": "object",
                "xml": {
                    "name": "ServiceTranslationResource"
                }
            },
            "SubserviceCollection": {
                "title": "SubserviceCollection",
                "properties": {
                    "data": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/SubserviceResource"
                        }
                    }
                },
                "type": "object"
            },
            "SubserviceResource": {
                "title": "SubserviceResource",
                "properties": {
                    "id": {
                        "$ref": "#/components/schemas/Subservice/properties/id"
                    },
                    "image": {
                        "$ref": "#/components/schemas/Subservice/properties/image"
                    },
                    "createdAt": {
                        "$ref": "#/components/schemas/Subservice/properties/created_at"
                    },
                    "updatedAt": {
                        "$ref": "#/components/schemas/Subservice/properties/updated_at"
                    },
                    "translations": {
                        "type": "array",
                        "items": {
                            "anyOf": [
                                {
                                    "$ref": "#/components/schemas/SubserviceTranslationResource"
                                }
                            ]
                        }
                    }
                },
                "type": "object",
                "xml": {
                    "name": "SubserviceResource"
                }
            },
...
            "SubserviceTranslationResource": {
                "title": "SubserviceTranslationResource",
                "properties": {
                    "id": {
                        "$ref": "#/components/schemas/SubserviceTranslation/properties/id"
                    },
                    "subserviceId": {
                        "$ref": "#/components/schemas/SubserviceTranslation/properties/subservice_id"
                    },
                    "locale": {
                        "$ref": "#/components/schemas/SubserviceTranslation/properties/locale"
                    },
                    "name": {
                        "$ref": "#/components/schemas/SubserviceTranslation/properties/name"
                    },
                    "content": {
                        "$ref": "#/components/schemas/SubserviceTranslation/properties/content"
                    }
                },
                "type": "object",
                "xml": {
                    "name": "SubserviceTranslationResource"
                }
            },
...
        },
        "securitySchemes": {
            "sanctum": {
                "type": "apiKey",
                "description": "Enter token in format (Bearer <token>)",
                "name": "Authorization",
                "in": "header"
            }
        }
    },
...
}